Update tests and validations to pass
I had to change the transfer request validations to make sure it let
the transfer go through if you were trying to transfer any and all URLs
that you own. This validation also broke the factories and some of the
standard tests, so those were updated to account for the new
validation.