fix transfer_request_spec tests
Three types of errors:

1. The app uses hidden inputs throughout, so add
a `js_make_all_inputs_visible` method
2. Some tests fail because the page isn't fully loaded with data,
so we add a wait_for_urls_page_load method
which makes sure "processing" isn't displayed
3. Modals. Sometimes we need to wait for the modal
to appear. Others we need to wait for it to be fully
dismissed. We add two helpers.

TODO: Flatten and simplify. Lots of nesting makes the test intent
difficult to decipher.  Also, there's a lot of waiting
which is making for slower tests (1m18s). Find
ways to speed this up.