Source apparition gem from github (#57)
The latest version of apparition gem can create a lot of noise with
`unknown keyword: :type:` warnings

This is fixed in <https://github.com/twalpole/apparition/pull/79>, but the patch hasn't been released yet. We change to github source in the interim.

Once patch is released, we should change back. See issue #58.