add route in dev and test modes for fake logins
With this, we can to `cy.login()` to quickly login a given user, and skip the normal auth sequence.

This is based on code from the rails cypress package for authentication. It's almost identical to how the Laravel cypress package does it.

This route is only available in dev and test modes.