update paper_trail and fix: db:migrate failing
This updates paper_trail gem to work with Rails 7.

Previously, the db migration was failing with the error:
```
wrong number of arguments (given 2, expected 1)
```

This seems to be cause by the change with keyword arguments from ruby 2 to ruby 3. Adding a doublesplat to spread the `versions_table_options` hash resolves things.