change admin urls table to vue format
I was on the fence about this one.

The admin urls table and the regular urls table have a bunch of shared javascript, which makes it difficult to update one without impacting the other.

After trying some updates to the erb template and the jquery, it seemed like I'd need to rework a bunch of old js and possibly risk breaking something on the non-admin side of things. So, I decided to switch tacts and just rework the page as a vue component like the collections and audit pages.

The downside of this approach is the urls page has more functionality: I needing to recreate the bulk actions, editing, deleting actions, modals etc.

So, there's a bit of code in this commit.