remove "Remove User" button
This turns out to be more complicated than it seems. To remove a url from a collection, we need to move it to a different collection. Who should own the URL? The admin then doing the removing?

If this is the case, it would probably be better to just have a list of admin collections as a dropdown under the “Owner” column, so they can see ALL of their colllections. It would make it clearer that it’s owned by them.

Deferring for now. Just removing the remove button.