test: Handle case where no meta tag with csrf token exists
In test mode, rails doesn't include a CSRF token in the headers because it doesn't bother checking for forgery while testing. So, when we initialize axios when RAILS_ENV=test and try to get a CSRF token, it causes an unhandled exception. Cypress fails when there are unhandled exceptions.  We just need a bit of optional chaining to avoid the exception.