fix url blurb buttons showing while closing
When a url is created, a section below it appears: what's called the `url-blurb` in the app. When the close button is clicked for this panel, the contents can overflow the container while the transition to close is happening.

This is because I removed the overflow-hidden class on the container so that the split-button dropdown wouldn't be cutoff.

This commit will add a the overflow-hidden back while the close is in progress, so that the contents won't appear.