fix: 500 status returned when creating collection
Cypress tests are failing when trying to create a new collection with the error:

```
Tried to load unspecified class: Time (Psych::DisallowedClass)
```

What's going on here?
Psych is a gem that serializes/deserializes objects into YAML for storage in the DB.  There's can be a security issue if any object is permitted to be stored, so Rails requires us to explicitly state which classes can be serialized into YAML.

Adding this config, resolves things:
```rb
config.active_record.yaml_column_permitted_classes = [Symbol, Date, Time]
```

**What's using YAML serialization in this app?**

I believe it's the PaperTrail gem, which we use for auditing changes to Zlinks and Collections. In the database, there's a `versions` table with a `object ` column that seems to have YAML data. Some of the data is date/time, which would require us to permit Date and Time explicitly.