remove admin announcements
In Z, admins can post global announcements for users. The intended functionality was to announce big updates or new features.

This functionlity relies on the [starburst gem](https://github.com/starburstgem/starburst). The gem is unmaintained. We forked it for the update to Rails 7.0.

When working on the update to Rails 7.1, it's apparent that our fork will need an update as the announcement tests are failing with errors referencing this gem. Perhaps minor, perhaps not.

Considering the admin announcements feature hasn't been used in years (if ever), it seems prudent to just remove the feature rather than maintaining it (and the startburst gem fork).

This PR removes the starbust gem and any admin announcement functionality.