fix: passenger error - base64 gem mismatch
After deploy, passenger shows error page. In the log:
```
You have already activated base64 0.1.1, but your Gemfile requires base64 0.2.0. Since base64 is a default gem, you can either remove your dependency on it or try updating to a newer version of bundler that supports base64 as a default gem. (Gem::LoadError)
```

We had the same error when updating from 7.0 to 7.1. A better solution is to update ruby, but this should work for now.