fix: `NoMethodError at /auth/saml/metadata`
When hitting the `/auth/saml/metadata` endpoint, the app throw a s no method error. The `metadata_signed` is set to true, a digest method is required for hashing the metadata before signature.

This sets the digest method to SHA256 and consolidates the shib_settings file and the omniauth.rb file